Various fixes for Visual Studio: - Disable Eigen alignment/vectorization to avoid weird access violation errors   in 32-bit builds - Update 'int' to 'std::int64_t' in a few last places - Various fixes to avoid integer conversion warnings - Suppress Visual Studio 'warning' that explicitly-initialized array member   variable (MatrixBase::_components) will in fact be value-initialized as per   the C++ standard - Fix name resolution for Zero within DoubleVectorBase

+++ OpenSolid/Core/BoundsFunction.hpp
@@ -43,19 +43,19 @@ namespace opensolid
     inline
     Interval
     BoundsFunction<int>::operator()(int value) const {
-        return Interval(value);
+        return Interval(double(value));
     }
 
     inline
     Interval
     BoundsFunction<unsigned>::operator()(unsigned value) const {
-        return Interval(value);
+        return Interval(double(value));
     }
 
     inline
     Interval
     BoundsFunction<std::int64_t>::operator()(std::int64_t value) const {
-        return Interval(value);
+        return Interval(double(value));
     }
 
     inline
